update web server (tcp server v2)

+++ WebServer/client.cpp
@@ -5,11 +5,11 @@
 #include "client.h"
 
 namespace webserver {
-    void client::set_id(int client_id) {
+    void client::set_id(unsigned long client_id) {
         id = client_id;
     }
 
-    int client::get_id() {
+    unsigned long client::get_id() {
         return id;
     }
 }
